Find native code in non-worklist modules
Formerly we only looked for native code in `worklist` modules.
This adds the new "external method extensions" (new methods
owned by the `worklist` but for functions owned by non-`worklist`
modules) and new specialization of external methods.
